cleanup: per-frame block lookup cache + clippy fixes
Perf: draw_viewport's `block_at` per visible row is the hot path in
long sessions (lots of OSC 133 marks accumulated). Adjacent rows almost
always belong to the same block; caching the last result drops a
typical frame from O(blocks × rows) to O(rows) for the common case.
Applied to all 4 viewport draw paths (CPU live, CPU smooth-scroll, GPU
live, GPU smooth-scroll).

Clippy:
- block_from_mark: drop redundant `as usize` cast.
- HUD section drag: drop dead-equal if (was `if to > from { to } else
  { to }` — both arms same).
- perform_key_equivalent: collapse outer `if let` arm into match guard.

Clippy is now clean. 783 workspace tests pass.

+++ crates/anvil-term/src/terminal.rs
@@ -630,7 +630,7 @@ impl Terminal {
         }
 
         let cursor_abs = self.evicted_lines + self.line_count().saturating_sub(self.rows())
-            + self.active_const().cur_y as usize;
+            + self.active_const().cur_y;
         let mut b = Block {
             command_line,
             command_start_col,
